requirements.txt
openai>=1.0.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
rank-bm25>=0.2.0
numpy>=1.24.0
matplotlib>=3.5.0
seaborn>=0.11.0
requests>=2.28.0
